pyproject.toml
[project]
name = "wiki-js-content-scrape"
version = "0.1.0"
description = "Scraper for wiki.js content pages"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "python-dotenv>=1.0",
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
